Move retry helper to base device
